+++ src/core/Table/Table.tsx
@@ -16,7 +16,6 @@ import {
   TableState,
   useFlexLayout,
   useFilters,
-  useMountedLayoutEffect,
   useRowSelect,
   useSortBy,
   useTable,
@@ -183,14 +182,12 @@ export const Table = <
   );
 
   // useRef prevents from rerendering when one of these callbacks changes
-  const onSelectRef = React.useRef(onSelect);
   const onBottomReachedRef = React.useRef(onBottomReached);
   const onRowInViewportRef = React.useRef(onRowInViewport);
   React.useEffect(() => {
-    onSelectRef.current = onSelect;
     onBottomReachedRef.current = onBottomReached;
     onRowInViewportRef.current = onRowInViewport;
-  }, [onBottomReached, onRowInViewport, onSelect]);
+  }, [onBottomReached, onRowInViewport]);
 
   const useSelectionHook = (hooks: Hooks<T>) => {
     if (!isSelectable) {
@@ -246,6 +243,24 @@ export const Table = <
     }
   };
 
+  const onSelectHandler = (
+    newState: TableState<T>,
+    instance?: TableInstance<T>,
+  ) => {
+    if (!instance?.rows.length) {
+      onSelect?.([], newState);
+      return;
+    }
+
+    const selectedData: T[] = [];
+    instance.rows.forEach((row) => {
+      if (newState.selectedRowIds[row.id]) {
+        selectedData.push(row.original);
+      }
+    });
+    onSelect?.(selectedData, newState);
+  };
+
   const tableStateReducer = (
     newState: TableState<T>,
     action: ActionType,
@@ -259,6 +274,12 @@ export const Table = <
       case TableActions.setFilter:
         onFilterHandler(newState, action, previousState, instance);
         break;
+      case TableActions.toggleRowSelected:
+      case TableActions.toggleAllRowsSelected:
+      case TableActions.toggleAllPageRowsSelected: {
+        onSelectHandler(newState, instance);
+        break;
+      }
       default:
         break;
     }
@@ -290,20 +311,10 @@ export const Table = <
     getTableBodyProps,
     prepareRow,
     data,
-    selectedFlatRows,
     state,
     allColumns,
     filteredFlatRows,
   } = instance;
-
-  useMountedLayoutEffect(() => {
-    if (isSelectable && selectedFlatRows) {
-      onSelectRef.current?.(
-        selectedFlatRows.map((row) => row.original),
-        state,
-      );
-    }
-  }, [selectedFlatRows, onSelectRef]);
 
   const ariaDataAttributes = Object.entries(rest).reduce(
     (result, [key, value]) => {